# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
...
  /api/job-seeker/resume/upload:
    post:
      summary: 'Upload resume'
      operationId: uploadResume
      description: "Uploads a resume file to the job seeker's profile without triggering AI analysis."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Resume uploaded successfully'
                  resume_url: 'https://example.com/storage/resumes/cv.pdf'
                properties:
                  message:
                    type: string
                    example: 'Resume uploaded successfully'
                  resume_url:
                    type: string
                    example: 'https://example.com/storage/resumes/cv.pdf'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors:
                    resume:
                      - 'The resume field is required.'
                properties:
                  errors:
                    type: object
                    properties:
                      resume:
                        type: array
                        example:
                          - 'The resume field is required.'
                        items:
                          type: string
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resume:
                  type: string
                  format: binary
                  description: 'PDF/DOC/DOCX file, max 5 MB.'
              required:
                - resume
...
  /api/job-seeker/jobs/search:
    get:
      summary: 'Search jobs'
      operationId: searchJobs
      description: 'Paginated search of active job posts for the authenticated job seeker.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  jobs:
                    data:
                      -
                        id: 664f1a2b3c4d5e6f7a8b9c0d
                        title: 'Senior React Developer'
                        company_name: 'Acme Corp'
                    current_page: 1
                    per_page: 15
                    total: 1
                    total_pages: 1
                    next_page: null
                    prev_page: null
                properties:
                  jobs:
                    type: object
                    properties:
                      data:
                        type: array
                        example:
                          -
                            id: 664f1a2b3c4d5e6f7a8b9c0d
                            title: 'Senior React Developer'
                            company_name: 'Acme Corp'
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: 664f1a2b3c4d5e6f7a8b9c0d
                            title:
                              type: string
                              example: 'Senior React Developer'
                            company_name:
                              type: string
                              example: 'Acme Corp'
                      current_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 15
                      total:
                        type: integer
                        example: 1
                      total_pages:
                        type: integer
                        example: 1
                      next_page:
                        type: string
                        example: null
                        nullable: true
                      prev_page:
                        type: string
                        example: null
                        nullable: true
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                keyword:
                  type: string
                  description: 'Search in title, description, company name. Max 100 chars.'
                  example: React
                  nullable: true
                location:
                  type: string
                  description: 'Partial location match. Max 100 chars.'
                  example: Beirut
                  nullable: true
                job_type:
                  type: string
                  description: 'One of: full_time, part_time, contract, freelance.'
                  example: full_time
                  nullable: true
                category:
                  type: string
                  description: ''
                  example: Engineering
                  nullable: true
                min_salary:
                  type: number
                  description: 'Minimum salary.'
                  example: 2000.0
                  nullable: true
...